finished writing the fishcounter sensor code
I faced a lot of issues.
For example, I could not declare the function in the header and define them in
.cpp files. I faced this before and I remember either fixing it or finding the reason.

I need to do this again.

millis probably resets every wakeup. So my debouncing logic has to change. This means the fishcounter code must be changed.

micros() causes exceptions. I need to figure why so I can possibly use it instead of millis().

'